backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9


# PostgreSQL async driver
asyncpg==0.29.0

PyMuPDF==1.24.10

sentence-transformers==3.1.1

chromadb==0.5.15

groq==0.11.0
httpx==0.27.0 #groq uses 

python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2


passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0